requirements.txt
colorama
Jinja2
GitPython
